backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.4
python-dotenv==1.0.0
motor==3.0.0
asyncpg==0.27.0
pymongo==4.5.0
postgres